Last working salamander working!
I made quick fixes, and now, by using the integration of a velocity
function, given a target "angular distance," the salamander is able to
turn left (if press 1), right (if press 2), and go back to straight walk
(if press 3).
Though, I needed to set the hinges at the limbs to limit 0 because
turning was causing inertia. Ultimately, that inertia made the legs jump
from the induced torque. So now they look fixed.
I will look at the ODE documentation to see how we can move only 1 body
of a hinge joint, so that we would be able to synchronize legs and body
segments without the "jumping" effect.
If you want to see the jumping effect, just change the
legMaxAngleDeformation global variable in the constructor of the
Salamander to your previous pi/16, and after a few turnings to
left/right, if you go back to walking straight, you will see the flaw.
See you!